% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_individual_return_rec_allele_matrix.r
\name{get_min_by_sample}
\alias{get_min_by_sample}
\title{Get the reads for the recessive allele for all genes in an individual for all samples}
\usage{
get_min_by_sample(
  ind,
  metaData = metaData,
  GTEX_allelecounts_A = GTEX_allelecounts_A,
  GTEX_allelecounts_B = GTEX_allelecounts_B
)
}
\arguments{
\item{ind}{individuals of interest}

\item{metaData}{the metadata which tells us which individuals belong to which columns in the allele data}

\item{GTEX_allelecounts_A}{allelecounts A}

\item{GTEX_allelecounts_B}{allelecounts B}
}
\value{
Function returns matrix of the recessive allele for each individual
}
\description{
Get the reads for the recessive allele for all genes in an individual for all samples
}
