% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slopes_by_tissue.r
\name{get_slopes_by_tissue}
\alias{get_slopes_by_tissue}
\title{Get the reads from the recessive allele for all genes in an individual for all samples}
\usage{
get_slopes_by_tissue(
  index,
  list_gtex_max = list_gtex_max,
  list_gtex_min = list_gtex_min
)
}
\arguments{
\item{index}{index used to run in parallel-- the element of the list in the list_gtex_max/list_gtex_min that is to be used}

\item{list_gtex_max}{max reads allele by gene}

\item{list_gtex_min}{min reads allele by gene}
}
\value{
Function returns matrix.
}
\description{
Get the reads from the recessive allele for all genes in an individual for all samples
}
