% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_dom_allele_matrix.r
\name{return_dom_allele_matrix}
\alias{return_dom_allele_matrix}
\title{Get the reads for the dominant allele for all genes in an individual for all samples}
\usage{
return_dom_allele_matrix(samples_oi_A, samples_oi_B)
}
\arguments{
\item{samples_oi_A}{dataframe of allele A counts to be passed in}

\item{samples_oi_B}{dataframe of allele B counts to be passed in}

\item{group_names}{groupnames of the data}
}
\value{
Function returns matrix.
}
\description{
Get the reads for the dominant allele for all genes in an individual for all samples
}
